+++ libs/render_delegate/native_rprim.cpp
@@ -35,6 +35,9 @@ HdArnoldNativeRprim::HdArnoldNativeRprim(
 void HdArnoldNativeRprim::Sync(
     HdSceneDelegate* sceneDelegate, HdRenderParam* renderParam, HdDirtyBits* dirtyBits, const TfToken& reprToken)
 {
+    if (!GetRenderDelegate()->CanUpdateScene())
+        return;
+ 
     TF_UNUSED(reprToken);
     HdArnoldRenderParamInterrupt param(renderParam);
     const auto& id = GetId();